third_party/WebKit/LayoutTests/platform/win/external/wpt/user-timing/test_user_timing_measure_exceptions-expected.txt

This is a testharness.js-based test.
PASS window.performance is defined 
PASS window.performance.measure("measure", "mark"), where "mark" is a non-existent mark,  threw an exception. 
PASS window.performance.measure("measure", "mark"), where "mark" is a non-existent mark, threw a SYNTAX_ERR exception. 
PASS window.performance.measure("measure", "mark", "responseEnd"), where "mark" is a non-existent mark, threw an exception. 
PASS window.performance.measure("measure", "mark", "responseEnd"), where "mark" is a non-existent mark, threw a SYNTAX_ERR exception. 
PASS window.performance.measure("measure", "navigationStart", "mark"), where "mark" is a non-existent mark, threw an exception. 
PASS window.performance.measure("measure", "navigationStart", "mark"), where "mark" is a non-existent mark, threw a SYNTAX_ERR exception. 
PASS window.performance.measure("measure", "mark", "mark"), where "mark" is a non-existent mark, threw an exception. 
PASS window.performance.measure("measure", "mark", "mark"), where "mark" is a non-existent mark, threw a SYNTAX_ERR exception. 
PASS window.performance.measure("measure", "unloadEventStart"), where "unloadEventStart" is a navigation timing attribute with a value of 0, threw an exception. 
PASS window.performance.measure("measure", "unloadEventStart"), where "unloadEventStart" is a navigation timing attribute with a value of 0, threw an INVALID_ACCESS_ERR exception. 
PASS window.performance.measure("measure", "unloadEventStart", "responseEnd"), where "unloadEventStart" is a navigation timing attribute with a value of 0, threw an exception. 
PASS window.performance.measure("measure", "unloadEventStart", "responseEnd"), where "unloadEventStart" is a navigation timing attribute with a value of 0, threw an INVALID_ACCESS_ERR exception. 
PASS window.performance.measure("measure", "navigationStart", "unloadEventStart"), where "unloadEventStart" is a navigation timing attribute with a value of 0, threw an exception. 
PASS window.performance.measure("measure", "navigationStart", "unloadEventStart"), where "unloadEventStart" is a navigation timing attribute with a value of 0, threw an INVALID_ACCESS_ERR exception. 
PASS window.performance.measure("measure", "unloadEventStart", "unloadEventStart"), where "unloadEventStart" is a navigation timing attribute with a value of 0, threw an exception. 
PASS window.performance.measure("measure", "unloadEventStart", "unloadEventStart"), where "unloadEventStart" is a navigation timing attribute with a value of 0, threw an INVALID_ACCESS_ERR exception. 
Harness: the test ran to completion.
